Add files via upload
Project Colour Changer - scope was to create js program to change the background colour of the web page when user clicks a button - cycles through 3 colours and displays the name of the current background colour - basic styling added to button, also added a show/hide button for a picture of a cat because....why not!

Project To Do List - scope was to add new tasks to list when user submits form and display tasks.  Allow user to mark tasks as complete and remove from list.  Some of the code was gained by research and using google which I then manipulated and styled accordingly.